requirements.txt
cryptography==44.0.2
docx==0.2.4
pandas==2.2.3
PyYAML==6.0.2
Requests==2.32.3
six==1.17.0
python-dotenv==1.1.0
openpyxl==3.1.2
